Refactor CRM API into single-action controllers
