--- src/Crm/Transport/Controller/Api/V1/CrmController.php
@@ -1,482 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Crm\Transport\Controller\Api\V1;
-
-use App\Crm\Application\Service\CompanyApplicationListService;
-use App\Crm\Application\Service\TaskListService;
-use App\Crm\Domain\Entity\Company;
-use App\Crm\Domain\Entity\Crm;
-use App\Crm\Domain\Entity\Project;
-use App\Crm\Domain\Entity\Sprint;
-use App\Crm\Domain\Entity\Task;
-use App\Crm\Domain\Entity\TaskRequest;
-use App\Crm\Domain\Enum\ProjectStatus;
-use App\Crm\Domain\Enum\SprintStatus;
-use App\Crm\Domain\Enum\TaskPriority;
-use App\Crm\Domain\Enum\TaskRequestStatus;
-use App\Crm\Domain\Enum\TaskStatus;
-use App\Crm\Infrastructure\Repository\CompanyRepository;
-use App\Crm\Infrastructure\Repository\CrmRepository;
-use App\Crm\Infrastructure\Repository\ProjectRepository;
-use App\Crm\Infrastructure\Repository\SprintRepository;
-use App\Crm\Infrastructure\Repository\TaskRepository;
-use App\Crm\Infrastructure\Repository\TaskRequestRepository;
-use App\General\Application\Message\EntityCreated;
-use App\General\Application\Message\EntityDeleted;
-use App\Platform\Domain\Entity\Application;
-use App\Platform\Domain\Enum\PlatformKey;
-use App\User\Domain\Entity\User;
-use DateTimeImmutable;
-use Doctrine\ORM\EntityManagerInterface;
-use OpenApi\Attributes as OA;
-use Symfony\Bundle\SecurityBundle\Security;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Attribute\AsController;
-use Symfony\Component\HttpKernel\Exception\HttpException;
-use Symfony\Component\Messenger\MessageBusInterface;
-use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
-
-#[AsController]
-#[OA\Tag(name: 'Crm')]
-#[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
-final readonly class CrmController
-{
-    public function __construct(
-        private CompanyRepository $companyRepository,
-        private ProjectRepository $projectRepository,
-        private TaskRepository $taskRepository,
-        private TaskRequestRepository $taskRequestRepository,
-        private SprintRepository $sprintRepository,
-        private CrmRepository $crmRepository,
-        private TaskListService $taskListService,
-        private CompanyApplicationListService $companyApplicationListService,
-        private EntityManagerInterface $entityManager,
-        private MessageBusInterface $messageBus,
-        private Security $security,
-    ) {
-    }
-
-    #[Route('/v1/crm/companies', methods: [Request::METHOD_GET])]
-    public function companies(): JsonResponse
-    {
-        $items = array_map(static fn (Company $company): array => [
-            'id' => $company->getId(),
-            'name' => $company->getName(),
-            'industry' => $company->getIndustry(),
-            'website' => $company->getWebsite(),
-            'contactEmail' => $company->getContactEmail(),
-            'phone' => $company->getPhone(),
-        ], $this->companyRepository->findBy([], [
-            'createdAt' => 'DESC',
-        ], 200));
-
-        return new JsonResponse([
-            'items' => $items,
-        ]);
-    }
-
-    #[Route('/v1/crm/companies', methods: [Request::METHOD_POST])]
-    #[OA\Post(summary: 'POST /v1/crm/companies', requestBody: new OA\RequestBody(required: true, content: new OA\JsonContent(required: ['payload'], properties: [
-        new OA\Property(property: 'payload', type: 'object', example: [
-            'value' => 'example',
-        ])], example: [
-            'payload' => [
-                'value' => 'example',
-            ],
-        ])), tags: ['Crm'], parameters: [], responses: [new OA\Response(response: 201, description: 'Success.'), new OA\Response(response: 400, description: 'Bad request.'), new OA\Response(response: 401, description: 'Unauthorized.'), new OA\Response(response: 404, description: 'Not found.'), new OA\Response(response: 422, description: 'Validation error.')])]
-    public function createCompany(Request $request): JsonResponse
-    {
-        $payload = (array)json_decode((string)$request->getContent(), true);
-        $company = new Company();
-        $company->setName((string)($payload['name'] ?? ''))
-            ->setIndustry(isset($payload['industry']) ? (string)$payload['industry'] : null)
-            ->setWebsite(isset($payload['website']) ? (string)$payload['website'] : null)
-            ->setContactEmail(isset($payload['contactEmail']) ? (string)$payload['contactEmail'] : null)
-            ->setPhone(isset($payload['phone']) ? (string)$payload['phone'] : null);
-        if (is_string($payload['crmId'] ?? null)) {
-            $company->setCrm($this->crmRepository->find($payload['crmId']));
-        }
-
-        $this->entityManager->persist($company);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityCreated('crm_company', $company->getId()));
-
-        return new JsonResponse([
-            'id' => $company->getId(),
-        ], JsonResponse::HTTP_CREATED);
-    }
-
-    #[Route('/v1/crm/companies/{id}', methods: [Request::METHOD_DELETE])]
-    public function deleteCompany(string $id): JsonResponse
-    {
-        $company = $this->companyRepository->find($id);
-        if (!$company instanceof Company) {
-            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
-        }
-
-        $this->entityManager->remove($company);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityDeleted('crm_company', $id));
-
-        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
-    }
-
-    #[Route('/v1/crm/projects', methods: [Request::METHOD_GET])]
-    public function projects(): JsonResponse
-    {
-        $items = array_map(static fn (Project $project): array => [
-            'id' => $project->getId(),
-            'name' => $project->getName(),
-            'companyId' => $project->getCompany()?->getId(),
-            'code' => $project->getCode(),
-            'status' => $project->getStatus()->value,
-            'startedAt' => $project->getStartedAt()?->format(DATE_ATOM),
-            'dueAt' => $project->getDueAt()?->format(DATE_ATOM),
-        ], $this->projectRepository->findBy([], [
-            'createdAt' => 'DESC',
-        ], 200));
-
-        return new JsonResponse([
-            'items' => $items,
-        ]);
-    }
-
-    #[Route('/v1/crm/projects', methods: [Request::METHOD_POST])]
-    #[OA\Post(summary: 'POST /v1/crm/projects', requestBody: new OA\RequestBody(required: true, content: new OA\JsonContent(required: ['payload'], properties: [
-        new OA\Property(property: 'payload', type: 'object', example: [
-            'value' => 'example',
-        ])], example: [
-            'payload' => [
-                'value' => 'example',
-            ],
-        ])), tags: ['Crm'], parameters: [], responses: [new OA\Response(response: 201, description: 'Success.'), new OA\Response(response: 400, description: 'Bad request.'), new OA\Response(response: 401, description: 'Unauthorized.'), new OA\Response(response: 404, description: 'Not found.'), new OA\Response(response: 422, description: 'Validation error.')])]
-    public function createProject(Request $request): JsonResponse
-    {
-        $payload = (array)json_decode((string)$request->getContent(), true);
-        $project = new Project();
-        $project->setName((string)($payload['name'] ?? ''))
-            ->setCode(isset($payload['code']) ? (string)$payload['code'] : null)
-            ->setDescription(isset($payload['description']) ? (string)$payload['description'] : null)
-            ->setStatus(ProjectStatus::tryFrom((string)($payload['status'] ?? '')) ?? ProjectStatus::PLANNED)
-            ->setStartedAt(isset($payload['startedAt']) ? new DateTimeImmutable((string)$payload['startedAt']) : null)
-            ->setDueAt(isset($payload['dueAt']) ? new DateTimeImmutable((string)$payload['dueAt']) : null);
-        if (is_string($payload['companyId'] ?? null)) {
-            $project->setCompany($this->companyRepository->find($payload['companyId']));
-        }
-
-        $this->entityManager->persist($project);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityCreated('crm_project', $project->getId()));
-
-        return new JsonResponse([
-            'id' => $project->getId(),
-        ], JsonResponse::HTTP_CREATED);
-    }
-
-    #[Route('/v1/crm/projects/{id}', methods: [Request::METHOD_DELETE])]
-    public function deleteProject(string $id): JsonResponse
-    {
-        $project = $this->projectRepository->find($id);
-        if (!$project instanceof Project) {
-            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
-        }
-
-        $this->entityManager->remove($project);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityDeleted('crm_project', $id));
-
-        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
-    }
-
-    #[Route('/v1/crm/tasks', methods: [Request::METHOD_GET])]
-    public function tasks(Request $request): JsonResponse
-    {
-        return new JsonResponse($this->taskListService->getList($request));
-    }
-
-    #[Route('/v1/crm/tasks', methods: [Request::METHOD_POST])]
-    #[OA\Post(summary: 'POST /v1/crm/tasks', requestBody: new OA\RequestBody(required: true, content: new OA\JsonContent(required: ['payload'], properties: [
-        new OA\Property(property: 'payload', type: 'object', example: [
-            'value' => 'example',
-        ])], example: [
-            'payload' => [
-                'value' => 'example',
-            ],
-        ])), tags: ['Crm'], parameters: [], responses: [new OA\Response(response: 201, description: 'Success.'), new OA\Response(response: 400, description: 'Bad request.'), new OA\Response(response: 401, description: 'Unauthorized.'), new OA\Response(response: 404, description: 'Not found.'), new OA\Response(response: 422, description: 'Validation error.')])]
-    public function createTask(Request $request): JsonResponse
-    {
-        $payload = (array)json_decode((string)$request->getContent(), true);
-        $task = new Task();
-        $task->setTitle((string)($payload['title'] ?? ''))
-            ->setDescription(isset($payload['description']) ? (string)$payload['description'] : null)
-            ->setStatus(TaskStatus::tryFrom((string)($payload['status'] ?? '')) ?? TaskStatus::TODO)
-            ->setPriority(TaskPriority::tryFrom((string)($payload['priority'] ?? '')) ?? TaskPriority::MEDIUM)
-            ->setDueAt(isset($payload['dueAt']) ? new DateTimeImmutable((string)$payload['dueAt']) : null)
-            ->setEstimatedHours(isset($payload['estimatedHours']) ? (float)$payload['estimatedHours'] : null);
-        if (is_string($payload['projectId'] ?? null)) {
-            $task->setProject($this->projectRepository->find($payload['projectId']));
-        }
-        if (is_string($payload['sprintId'] ?? null)) {
-            $task->setSprint($this->sprintRepository->find($payload['sprintId']));
-        }
-
-        $this->entityManager->persist($task);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityCreated('crm_task', $task->getId()));
-
-        return new JsonResponse([
-            'id' => $task->getId(),
-        ], JsonResponse::HTTP_CREATED);
-    }
-
-    #[Route('/v1/crm/tasks/{id}', methods: [Request::METHOD_DELETE])]
-    public function deleteTask(string $id): JsonResponse
-    {
-        $task = $this->taskRepository->find($id);
-        if (!$task instanceof Task) {
-            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
-        }
-
-        $this->entityManager->remove($task);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityDeleted('crm_task', $id));
-
-        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
-    }
-
-    #[Route('/v1/crm/task-requests', methods: [Request::METHOD_GET])]
-    public function taskRequests(): JsonResponse
-    {
-        $items = array_map(static fn (TaskRequest $taskRequest): array => [
-            'id' => $taskRequest->getId(),
-            'title' => $taskRequest->getTitle(),
-            'status' => $taskRequest->getStatus()->value,
-            'requestedAt' => $taskRequest->getRequestedAt()->format(DATE_ATOM),
-            'resolvedAt' => $taskRequest->getResolvedAt()?->format(DATE_ATOM),
-            'taskId' => $taskRequest->getTask()?->getId(),
-        ], $this->taskRequestRepository->findBy([], [
-            'createdAt' => 'DESC',
-        ], 200));
-
-        return new JsonResponse([
-            'items' => $items,
-        ]);
-    }
-
-    #[Route('/v1/crm/task-requests', methods: [Request::METHOD_POST])]
-    #[OA\Post(summary: 'POST /v1/crm/task-requests', requestBody: new OA\RequestBody(required: true, content: new OA\JsonContent(required: ['payload'], properties: [
-        new OA\Property(property: 'payload', type: 'object', example: [
-            'value' => 'example',
-        ])], example: [
-            'payload' => [
-                'value' => 'example',
-            ],
-        ])), tags: ['Crm'], parameters: [], responses: [new OA\Response(response: 201, description: 'Success.'), new OA\Response(response: 400, description: 'Bad request.'), new OA\Response(response: 401, description: 'Unauthorized.'), new OA\Response(response: 404, description: 'Not found.'), new OA\Response(response: 422, description: 'Validation error.')])]
-    public function createTaskRequest(Request $request): JsonResponse
-    {
-        $payload = (array)json_decode((string)$request->getContent(), true);
-        $taskRequest = new TaskRequest();
-        $taskRequest->setTitle((string)($payload['title'] ?? ''))
-            ->setDescription(isset($payload['description']) ? (string)$payload['description'] : null)
-            ->setStatus(TaskRequestStatus::tryFrom((string)($payload['status'] ?? '')) ?? TaskRequestStatus::PENDING)
-            ->setResolvedAt(isset($payload['resolvedAt']) ? new DateTimeImmutable((string)$payload['resolvedAt']) : null);
-        if (is_string($payload['taskId'] ?? null)) {
-            $taskRequest->setTask($this->taskRepository->find($payload['taskId']));
-        }
-
-        $this->entityManager->persist($taskRequest);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityCreated('crm_task_request', $taskRequest->getId()));
-
-        return new JsonResponse([
-            'id' => $taskRequest->getId(),
-        ], JsonResponse::HTTP_CREATED);
-    }
-
-    #[Route('/v1/crm/task-requests/{id}', methods: [Request::METHOD_DELETE])]
-    public function deleteTaskRequest(string $id): JsonResponse
-    {
-        $taskRequest = $this->taskRequestRepository->find($id);
-        if (!$taskRequest instanceof TaskRequest) {
-            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
-        }
-
-        $this->entityManager->remove($taskRequest);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityDeleted('crm_task_request', $id));
-
-        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
-    }
-
-    #[Route('/v1/crm/sprints', methods: [Request::METHOD_GET])]
-    public function sprints(): JsonResponse
-    {
-        $items = array_map(static fn (Sprint $sprint): array => [
-            'id' => $sprint->getId(),
-            'name' => $sprint->getName(),
-            'projectId' => $sprint->getProject()?->getId(),
-            'status' => $sprint->getStatus()->value,
-            'startDate' => $sprint->getStartDate()?->format('Y-m-d'),
-            'endDate' => $sprint->getEndDate()?->format('Y-m-d'),
-        ], $this->sprintRepository->findBy([], [
-            'createdAt' => 'DESC',
-        ], 200));
-
-        return new JsonResponse([
-            'items' => $items,
-        ]);
-    }
-
-    #[Route('/v1/crm/sprints', methods: [Request::METHOD_POST])]
-    #[OA\Post(summary: 'POST /v1/crm/sprints', requestBody: new OA\RequestBody(required: true, content: new OA\JsonContent(required: ['payload'], properties: [
-        new OA\Property(property: 'payload', type: 'object', example: [
-            'value' => 'example',
-        ])], example: [
-            'payload' => [
-                'value' => 'example',
-            ],
-        ])), tags: ['Crm'], parameters: [], responses: [new OA\Response(response: 201, description: 'Success.'), new OA\Response(response: 400, description: 'Bad request.'), new OA\Response(response: 401, description: 'Unauthorized.'), new OA\Response(response: 404, description: 'Not found.'), new OA\Response(response: 422, description: 'Validation error.')])]
-    public function createSprint(Request $request): JsonResponse
-    {
-        $payload = (array)json_decode((string)$request->getContent(), true);
-        $sprint = new Sprint();
-        $sprint->setName((string)($payload['name'] ?? ''))
-            ->setGoal(isset($payload['goal']) ? (string)$payload['goal'] : null)
-            ->setStatus(SprintStatus::tryFrom((string)($payload['status'] ?? '')) ?? SprintStatus::PLANNED)
-            ->setStartDate(isset($payload['startDate']) ? new DateTimeImmutable((string)$payload['startDate']) : null)
-            ->setEndDate(isset($payload['endDate']) ? new DateTimeImmutable((string)$payload['endDate']) : null);
-        if (is_string($payload['projectId'] ?? null)) {
-            $sprint->setProject($this->projectRepository->find($payload['projectId']));
-        }
-
-        $this->entityManager->persist($sprint);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityCreated('crm_sprint', $sprint->getId()));
-
-        return new JsonResponse([
-            'id' => $sprint->getId(),
-        ], JsonResponse::HTTP_CREATED);
-    }
-
-    #[Route('/v1/crm/sprints/{id}', methods: [Request::METHOD_DELETE])]
-    public function deleteSprint(string $id): JsonResponse
-    {
-        $sprint = $this->sprintRepository->find($id);
-        if (!$sprint instanceof Sprint) {
-            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
-        }
-
-        $this->entityManager->remove($sprint);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityDeleted('crm_sprint', $id));
-
-        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
-    }
-
-    #[Route('/v1/crm/applications/{applicationSlug}/companies', methods: [Request::METHOD_GET])]
-    #[OA\Parameter(name: 'applicationSlug', in: 'path', required: true, schema: new OA\Schema(type: 'string'), example: 'crm-sales-hub')]
-    #[OA\Response(response: 200, description: 'Companies list scoped to CRM application.')]
-    public function companiesByApplication(string $applicationSlug, Request $request): JsonResponse
-    {
-        $crm = $this->resolveOrCreateCrmByApplicationSlug($applicationSlug);
-
-        return new JsonResponse($this->companyApplicationListService->getList($request, $applicationSlug, $crm));
-    }
-
-    #[Route('/v1/crm/applications/{applicationSlug}/companies', methods: [Request::METHOD_POST])]
-    #[OA\Post(summary: 'POST /v1/crm/applications/{applicationSlug}/companies', tags: ['Crm'], parameters: [new OA\Parameter(name: 'applicationSlug', in: 'path', required: true, schema: new OA\Schema(type: 'string'))], responses: [new OA\Response(response: 201, description: 'Success.'), new OA\Response(response: 400, description: 'Bad request.'), new OA\Response(response: 401, description: 'Unauthorized.'), new OA\Response(response: 404, description: 'Not found.'), new OA\Response(response: 422, description: 'Validation error.')])]
-    #[OA\Parameter(name: 'applicationSlug', in: 'path', required: true, schema: new OA\Schema(type: 'string'), example: 'crm-sales-hub')]
-    #[OA\RequestBody(
-        required: true,
-        content: new OA\JsonContent(
-            required: ['name'],
-            properties: [
-                new OA\Property(property: 'name', type: 'string', minLength: 1, example: 'Acme Europe'),
-            ],
-            type: 'object',
-            example: [
-                'name' => 'Acme Europe',
-            ],
-        )
-    )]
-    public function createCompanyByApplication(string $applicationSlug, Request $request): JsonResponse
-    {
-        $crm = $this->resolveOrCreateCrmByApplicationSlug($applicationSlug);
-        $payload = (array)json_decode((string)$request->getContent(), true);
-
-        $company = (new Company())
-            ->setCrm($crm)
-            ->setName((string)($payload['name'] ?? ''))
-            ->setIndustry(isset($payload['industry']) ? (string)$payload['industry'] : null)
-            ->setWebsite(isset($payload['website']) ? (string)$payload['website'] : null)
-            ->setContactEmail(isset($payload['contactEmail']) ? (string)$payload['contactEmail'] : null)
-            ->setPhone(isset($payload['phone']) ? (string)$payload['phone'] : null);
-
-        $this->entityManager->persist($company);
-        $this->entityManager->flush();
-        $this->messageBus->dispatch(new EntityCreated('crm_company', $company->getId(), context: [
-            'applicationSlug' => $applicationSlug,
-        ]));
-
-        return new JsonResponse([
-            'id' => $company->getId(),
-            'crmId' => $crm->getId(),
-            'applicationSlug' => $applicationSlug,
-        ], JsonResponse::HTTP_CREATED);
-    }
-
-
-    #[Route('/v1/crm/dashboard', methods: [Request::METHOD_GET])]
-    public function dashboard(): JsonResponse
-    {
-        return new JsonResponse([
-            'companies' => $this->companyRepository->count([]),
-            'projects' => $this->projectRepository->count([]),
-            'tasks' => $this->taskRepository->count([]),
-            'taskRequests' => [
-                TaskRequestStatus::PENDING->value => $this->taskRequestRepository->count(['status' => TaskRequestStatus::PENDING]),
-                TaskRequestStatus::APPROVED->value => $this->taskRequestRepository->count(['status' => TaskRequestStatus::APPROVED]),
-                TaskRequestStatus::REJECTED->value => $this->taskRequestRepository->count(['status' => TaskRequestStatus::REJECTED]),
-            ],
-        ]);
-    }
-
-    private function resolveOrCreateCrmByApplicationSlug(string $applicationSlug): Crm
-    {
-        $crm = $this->crmRepository->findOneByApplicationSlug($applicationSlug);
-        if ($crm instanceof Crm) {
-            $this->assertApplicationAccess($crm->getApplication(), PlatformKey::CRM);
-
-            return $crm;
-        }
-
-        $application = $this->entityManager->getRepository(Application::class)->findOneBy([
-            'slug' => $applicationSlug,
-        ]);
-        if (!$application instanceof Application) {
-            throw new HttpException(JsonResponse::HTTP_NOT_FOUND, 'Unknown "applicationSlug".');
-        }
-
-        $this->assertApplicationAccess($application, PlatformKey::CRM);
-
-        throw new HttpException(JsonResponse::HTTP_NOT_FOUND, 'CRM root entity not found for this application.');
-
-    }
-
-    private function assertApplicationAccess(?Application $application, PlatformKey $platformKey): void
-    {
-        if (!$application instanceof Application || $application->getPlatform()?->getPlatformKey() !== $platformKey) {
-            throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Invalid "applicationSlug" for the requested platform.');
-        }
-
-        $user = $this->security->getUser();
-        if (!$user instanceof User || $application->getUser()?->getId() !== $user->getId()) {
-            throw new HttpException(JsonResponse::HTTP_FORBIDDEN, 'You cannot access this application scope.');
-        }
-    }
-}
