+++ src/Crm/Transport/Controller/Api/V1/Project/DeleteProjectController.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Crm\Transport\Controller\Api\V1\Project;
+
+use App\Crm\Domain\Entity\Project;
+use App\Crm\Infrastructure\Repository\ProjectRepository;
+use App\General\Application\Message\EntityDeleted;
+use Doctrine\ORM\EntityManagerInterface;
+use OpenApi\Attributes as OA;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Attribute\AsController;
+use Symfony\Component\Messenger\MessageBusInterface;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+#[AsController]
+#[OA\Tag(name: 'Crm')]
+#[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
+final readonly class DeleteProjectController
+{
+    public function __construct(
+        private ProjectRepository $projectRepository,
+        private EntityManagerInterface $entityManager,
+        private MessageBusInterface $messageBus,
+    ) {
+    }
+
+    #[Route('/v1/crm/projects/{id}', methods: [Request::METHOD_DELETE])]
+    public function __invoke(string $id): JsonResponse
+    {
+        $project = $this->projectRepository->find($id);
+        if (!$project instanceof Project) {
+            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        $this->entityManager->remove($project);
+        $this->entityManager->flush();
+        $this->messageBus->dispatch(new EntityDeleted('crm_project', $id));
+
+        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
+    }
+}
