+++ src/Crm/Transport/Controller/Api/V1/TaskRequest/DeleteTaskRequestController.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Crm\Transport\Controller\Api\V1\TaskRequest;
+
+use App\Crm\Domain\Entity\TaskRequest;
+use App\Crm\Infrastructure\Repository\TaskRequestRepository;
+use App\General\Application\Message\EntityDeleted;
+use Doctrine\ORM\EntityManagerInterface;
+use OpenApi\Attributes as OA;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Attribute\AsController;
+use Symfony\Component\Messenger\MessageBusInterface;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+#[AsController]
+#[OA\Tag(name: 'Crm')]
+#[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
+final readonly class DeleteTaskRequestController
+{
+    public function __construct(
+        private TaskRequestRepository $taskRequestRepository,
+        private EntityManagerInterface $entityManager,
+        private MessageBusInterface $messageBus,
+    ) {
+    }
+
+    #[Route('/v1/crm/task-requests/{id}', methods: [Request::METHOD_DELETE])]
+    public function __invoke(string $id): JsonResponse
+    {
+        $taskRequest = $this->taskRequestRepository->find($id);
+        if (!$taskRequest instanceof TaskRequest) {
+            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        $this->entityManager->remove($taskRequest);
+        $this->entityManager->flush();
+        $this->messageBus->dispatch(new EntityDeleted('crm_task_request', $id));
+
+        return new JsonResponse(status: JsonResponse::HTTP_NO_CONTENT);
+    }
+}
